refactor(agentos): cross-daemon services bucket ai/services/{graph,ingestion}/ (#11846)
Authored by Claude Opus 4.7 (Claude Code). Session 5572d9a5-558d-4bea-b416-e31496c289c4.

FAIR-band: in-band [verify @ merge-gate]

Resolves #11846

Sub 10 of Epic #11831 — moves the 10 cross-daemon services out of
`ai/daemons/services/` (cleared by Subs 8 + 9) into thematic
`ai/services/{graph,ingestion}/` folders mirroring the existing
`ai/services/{github-workflow,gitlab-workflow,knowledge-base,memory-core,neural-link}/`
buckets. Empty `ai/daemons/services/` folder removed.

Source moves to ai/services/graph/ (6):
- GapInferenceEngine.mjs + GoldenPathSynthesizer.mjs + GraphMaintenanceService.mjs
- LazyEdgeDrainer.mjs + SemanticGraphExtractor.mjs + TopologyInferenceEngine.mjs

Source moves to ai/services/ingestion/ (4):
- ConceptDiscoveryService.mjs + ConceptIngestor.mjs + IssueIngestor.mjs
- MemorySessionIngestor.mjs

Test moves mirror source structure (7 files).

Depth preserved (both old `ai/daemons/services/X` and new
`ai/services/{graph,ingestion}/X` are 2-deep under `ai/`), so consumer
import rewrites are pure substring substitutions `daemons/services/X.mjs`
→ `services/{folder}/X.mjs` regardless of consumer location. 22 consumer
files updated.

Manual fixes for relative-path consumers whose strings DIDN'T contain the
literal `daemons/` segment:
- Orchestrator.mjs:32 — `'../services/GoldenPathSynthesizer.mjs'` (resolves
  to old via `ai/daemons/services/`) → `'../../services/graph/...'`
- DreamService.mjs:14-23 — 9× `'../../services/X.mjs'` pattern → `'../../../services/{folder}/...'`

V-B-A from Sub 9 (#11845) lesson: my regex-only sweep missed paths that
resolve to the moved location but don't contain the explicit `daemons/`
segment. Two-stage sweep (substring rewrite + grep for proximity-relative
patterns) caught both classes here.

Local validation:
- Moved-spec subsuite (graph/ingestion): 61/61 pass
- Orchestrator suite: 119/119 pass (DreamService spec required individual rerun
  due to parallel-worker flake; passes deterministically)
- Touched consumer specs (5): 130/130 pass
- Full unit suite: 2011 pass / 19 fail / 2 skip / 10 did-not-run — all 19
  failures are pre-existing flakes (Authorization, checkAllAgentIdle,
  checkSunsetted, QueryReRanker, SessionSummarization, TextEmbeddingService.retry,
  Grid family, vdom/table); verified zero references to moved files via grep.

Per @tobiu "you and me" instruction: no peer-review broadcast on this lane.
