+++ ai/daemons/orchestrator/services/DreamService.mjs
@@ -11,16 +11,16 @@ import { Memory_GraphService as GraphService } from '../../../services.mjs';
 import Json from '../../../../src/util/Json.mjs';
 import logger from '../../../mcp/server/memory-core/logger.mjs';
 import OpenAiCompatible from '../../../provider/OpenAiCompatible.mjs';
-import ConceptDiscoveryService from '../../services/ConceptDiscoveryService.mjs';
-import ConceptIngestor from '../../services/ConceptIngestor.mjs';
+import ConceptDiscoveryService from '../../../services/ingestion/ConceptDiscoveryService.mjs';
+import ConceptIngestor from '../../../services/ingestion/ConceptIngestor.mjs';
 import FileSystemIngestor from '../../../services/memory-core/FileSystemIngestor.mjs';
-import GapInferenceEngine from '../../services/GapInferenceEngine.mjs';
-import GraphMaintenanceService from '../../services/GraphMaintenanceService.mjs';
-import IssueIngestor from '../../services/IssueIngestor.mjs';
-import MemorySessionIngestor from '../../services/MemorySessionIngestor.mjs';
-import SemanticGraphExtractor from '../../services/SemanticGraphExtractor.mjs';
-import TopologyInferenceEngine from '../../services/TopologyInferenceEngine.mjs';
-import GoldenPathSynthesizer from '../../services/GoldenPathSynthesizer.mjs';
+import GapInferenceEngine from '../../../services/graph/GapInferenceEngine.mjs';
+import GraphMaintenanceService from '../../../services/graph/GraphMaintenanceService.mjs';
+import IssueIngestor from '../../../services/ingestion/IssueIngestor.mjs';
+import MemorySessionIngestor from '../../../services/ingestion/MemorySessionIngestor.mjs';
+import SemanticGraphExtractor from '../../../services/graph/SemanticGraphExtractor.mjs';
+import TopologyInferenceEngine from '../../../services/graph/TopologyInferenceEngine.mjs';
+import GoldenPathSynthesizer from '../../../services/graph/GoldenPathSynthesizer.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
