+++ ai/services/github-workflow/SyncService.mjs
@@ -208,7 +208,7 @@ class SyncService extends Base {
             // Dynamic import rationale: `IssueIngestor` depends on `GraphService` and `StorageRouter` (SQLite/ChromaDB).
             // Dynamically importing it here prevents the `github-workflow` MCP server from loading heavy database
             // dependencies or crashing on boot if the `memory-core` DB is locked, maintaining strict process boundary isolation.
-            const IssueIngestor = (await import('../../daemons/services/IssueIngestor.mjs')).default;
+            const IssueIngestor = (await import('../../services/ingestion/IssueIngestor.mjs')).default;
             await IssueIngestor.ingestIssueStates();
             await IssueIngestor.ingestDiscussionStates();
             await IssueIngestor.ingestPullRequestFeedback();
