+++ ai/services/graph/GoldenPathSynthesizer.mjs
