+++ buildScripts/ai/runSandman.mjs
@@ -4,7 +4,7 @@ import InstanceManager from '../../src/manager/Instance.mjs';
 import Memory_Config from '../../ai/mcp/server/memory-core/config.mjs';
 import Memory_Service from '../../ai/services/memory-core/MemoryService.mjs';
 import DreamService from '../../ai/daemons/orchestrator/services/DreamService.mjs';
-import GoldenPathSynthesizer from '../../ai/daemons/services/GoldenPathSynthesizer.mjs';
+import GoldenPathSynthesizer from '../../ai/services/graph/GoldenPathSynthesizer.mjs';
 import LifecycleService from '../../ai/services/memory-core/lifecycle/SystemLifecycleService.mjs';
 import InferenceLifecycleService from '../../ai/services/memory-core/lifecycle/InferenceLifecycleService.mjs';
 import GraphService from '../../ai/services/memory-core/GraphService.mjs';
